Move MessageListWidget to feature widget module
