+++ app/common/src/main/java/com/fsck/k9/widget/list/KoinModule.kt
@@ -1,6 +1,6 @@
 package com.fsck.k9.widget.list
 
-import app.k9mail.ui.widget.list.MessageListWidgetConfig
+import app.k9mail.feature.widget.message.list.MessageListWidgetConfig
 import org.koin.dsl.module
 
 val messageListWidgetConfigModule = module {
