+++ app/common/src/main/java/com/fsck/k9/widget/list/MessageListWidgetProvider.kt
@@ -1,8 +1,8 @@
 package com.fsck.k9.widget.list
 
-import app.k9mail.ui.widget.list.MessageListWidgetConfig
+import app.k9mail.feature.widget.message.list.MessageListWidgetConfig
 
-class MessageListWidgetProvider : app.k9mail.ui.widget.list.MessageListWidgetProvider()
+class MessageListWidgetProvider : app.k9mail.feature.widget.message.list.MessageListWidgetProvider()
 
 internal class K9MessageListWidgetConfig : MessageListWidgetConfig {
     override val providerClass = MessageListWidgetProvider::class.java
