+++ feature/widget/message-list/build.gradle.kts
@@ -1,0 +1,14 @@
+plugins {
+    id(ThunderbirdPlugins.Library.android)
+}
+
+dependencies {
+    implementation(projects.app.ui.legacy)
+    implementation(projects.app.core)
+
+    implementation(libs.timber)
+}
+
+android {
+    namespace = "app.k9mail.feature.widget.message.list"
+}
