+++ feature/widget/message-list/src/main/AndroidManifest.xml
