+++ feature/widget/message-list/src/main/kotlin/app/k9mail/feature/widget/message/list/KoinModule.kt
@@ -1,4 +1,4 @@
-package app.k9mail.ui.widget.list
+package app.k9mail.feature.widget.message.list
 
 import org.koin.dsl.module
