+++ feature/widget/message-list/src/main/kotlin/app/k9mail/feature/widget/message/list/MessageListConfig.kt
@@ -1,4 +1,4 @@
-package app.k9mail.ui.widget.list
+package app.k9mail.feature.widget.message.list
 
 import com.fsck.k9.Account.SortType
 import com.fsck.k9.search.LocalSearch
