+++ feature/widget/message-list/src/main/res/drawable-xxhdpi/message_list_widget_preview.png
