+++ feature/widget/message-list/src/main/res/layout/message_list_widget_list_item_loading.xml
