+++ feature/widget/message-list/src/main/res/layout/message_list_widget_loading.xml
