+++ feature/widget/message-list/src/main/res/values/colors.xml
