+++ feature/widget/message-list/src/main/res/xml/message_list_widget_info.xml
